+++ apsg/tensors.py
@@ -174,6 +174,27 @@ class DefGrad(np.ndarray):
         _, V = polar(self, 'left')
         return DefGrad(V)
 
+    @property
+    def axisangle(self):
+        """Return rotation part of ``DefGrad`` axis, angle tuple."""
+        from scipy.linalg import polar
+        R, _ = polar(self)
+        w, W = np.linalg.eig(R.T)
+        i = np.where(abs(np.real(w) - 1.0) < 1e-8)[0]
+        if not len(i):
+            raise ValueError("no unit eigenvector corresponding to eigenvalue 1")
+        axis = Vec3(np.real(W[:, i[-1]]).squeeze())
+        # rotation angle depending on direction
+        cosa = (np.trace(R) - 1.0) / 2.0
+        if abs(axis[2]) > 1e-8:
+            sina = (R[1, 0] + (cosa-1.0)*axis[0]*axis[1]) / axis[2]
+        elif abs(axis[1]) > 1e-8:
+            sina = (R[0, 2] + (cosa-1.0)*axis[0]*axis[2]) / axis[1]
+        else:
+            sina = (R[2, 1] + (cosa-1.0)*axis[1]*axis[2]) / axis[0]
+        angle = np.rad2deg(np.arctan2(sina, cosa))
+        return axis, angle
+
 
 class VelGrad(np.ndarray):
     """``VelGrad`` store velocity gradient tensor derived from numpy.ndarray